Fix grammatical error in the first section of the tutorial introduction
The comparison made with frameworks in a plural form, so we need to add the "s" at the end of framework.

The words "also" and "as well" in the same declaration is redundant, so we need to remove the add-on marker "as well" for simplicity.

The word "also" might be understood as connecting to the word frameworks, implying other frameworks also run in the browser, which is the opposite of what we are trying to communicate. So I'm suggesting we remove it, keeping the sentence direct.